requirements.txt
wget>=3.2
h5py>=2.9.0
numpy>=1.16.4
matplotlib>=3.0.3
scipy>=1.2.1
Pillow>=6.1.0
typing>=3.7.4
torch>=1.0.1
torchvision>=0.3.0
click>=7.0
pandas
segmentation_models_pytorch
albumentations
